+++ docs/graphics/index.js
@@ -1,12 +1,14 @@
 // @flow strict
 import AvatarGroup from './general/AvatarGroup.svg';
-import IconButton from './general/IconButton.svg';
 import Button from './general/Button.svg';
+import Icon from './general/Icon.svg';
+import Dropdown from './general/Dropdown.svg';
 
 const illustrations = Object.freeze({
   'avatar-group': AvatarGroup,
-  'icon-button': IconButton,
   'button': Button,
+  'dropdown': Dropdown,
+  'icon': Icon,
 });
 
 export type IllustrationTypes = $Keys<typeof illustrations>;